[CHORE] project model 분리
